10_Gamification_App/requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.54
langchain_google_genai==2.0.0
google-generativeai==0.7.0
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
# Change protobuf to a version that satisfies streamlit
protobuf>=3.20,<6.0
python-dotenv==1.1.0
streamlit==1.44.1
faiss-cpu
